Hotfix 4: Seasons and SeasonDAO
Fixed all attributes and constructors, "contracts" class included.
Fixed bugs on frmSeasons and made it functional so that it can read and select data from the seasons, its gps and its teams and its contracts
Now we can generate random seasons, only taking into account the year of it and the min-max values of the teams and gps to participate.